[pwa] Reorganize /district/ route
