+++ pwa/app/routeTree.gen.ts
@@ -34,7 +34,6 @@ import { Route as LocalDebugRouteImport } from './routes/local.debug'
 import { Route as MatchMatchKeyRouteImport } from './routes/match.$matchKey'
 import { Route as TeamsChar123PgNumChar125RouteImport } from './routes/teams.{-$pgNum}'
 import { Route as DistrictDistrictAbbreviationInsightsRouteImport } from './routes/district.$districtAbbreviation.insights'
-import { Route as DistrictDistrictAbbreviationStatsRouteImport } from './routes/district.$districtAbbreviation.stats'
 import { Route as DistrictDistrictAbbreviationChar123YearChar125RouteImport } from './routes/district.$districtAbbreviation.{-$year}'
 import { Route as EventsDistrictAbbreviationYearRouteImport } from './routes/events.$districtAbbreviation.$year'
 import { Route as SuggestReviewIndexRouteImport } from './routes/suggest.review.index'
@@ -43,7 +42,6 @@ import { Route as SuggestTeamMediaRouteImport } from './routes/suggest.team.medi
 import { Route as TeamTeamNumberHistoryRouteImport } from './routes/team.$teamNumber.history'
 import { Route as TeamTeamNumberStatsRouteImport } from './routes/team.$teamNumber.stats'
 import { Route as TeamTeamNumberChar123YearChar125RouteImport } from './routes/team.$teamNumber.{-$year}'
-import { Route as DistrictDistrictAbbreviationChampsYearRouteImport } from './routes/district.$districtAbbreviation.champs.$year'
 
 const IndexRoute = IndexRouteImport.update({
   id: '/',
@@ -175,12 +173,6 @@ const DistrictDistrictAbbreviationInsightsRoute =
     path: '/district/$districtAbbreviation/insights',
     getParentRoute: () => rootRouteImport,
   } as any)
-const DistrictDistrictAbbreviationStatsRoute =
-  DistrictDistrictAbbreviationStatsRouteImport.update({
-    id: '/district/$districtAbbreviation/stats',
-    path: '/district/$districtAbbreviation/stats',
-    getParentRoute: () => rootRouteImport,
-  } as any)
 const DistrictDistrictAbbreviationChar123YearChar125Route =
   DistrictDistrictAbbreviationChar123YearChar125RouteImport.update({
     id: '/district/$districtAbbreviation/{-$year}',
@@ -225,12 +217,6 @@ const TeamTeamNumberChar123YearChar125Route =
     path: '/team/$teamNumber/{-$year}',
     getParentRoute: () => rootRouteImport,
   } as any)
-const DistrictDistrictAbbreviationChampsYearRoute =
-  DistrictDistrictAbbreviationChampsYearRouteImport.update({
-    id: '/district/$districtAbbreviation/champs/$year',
-    path: '/district/$districtAbbreviation/champs/$year',
-    getParentRoute: () => rootRouteImport,
-  } as any)
 
 export interface FileRoutesByFullPath {
   '/': typeof IndexRoute
@@ -258,7 +244,6 @@ export interface FileRoutesByFullPath {
   '/teams/{-$pgNum}': typeof TeamsChar123PgNumChar125Route
   '/account/': typeof AccountIndexRoute
   '/district/$districtAbbreviation/insights': typeof DistrictDistrictAbbreviationInsightsRoute
-  '/district/$districtAbbreviation/stats': typeof DistrictDistrictAbbreviationStatsRoute
   '/district/$districtAbbreviation/{-$year}': typeof DistrictDistrictAbbreviationChar123YearChar125Route
   '/events/$districtAbbreviation/$year': typeof EventsDistrictAbbreviationYearRoute
   '/suggest/review/$suggestionType': typeof SuggestReviewSuggestionTypeRoute
@@ -267,7 +252,6 @@ export interface FileRoutesByFullPath {
   '/team/$teamNumber/stats': typeof TeamTeamNumberStatsRoute
   '/team/$teamNumber/{-$year}': typeof TeamTeamNumberChar123YearChar125Route
   '/suggest/review/': typeof SuggestReviewIndexRoute
-  '/district/$districtAbbreviation/champs/$year': typeof DistrictDistrictAbbreviationChampsYearRoute
 }
 export interface FileRoutesByTo {
   '/': typeof IndexRoute
@@ -295,7 +279,6 @@ export interface FileRoutesByTo {
   '/teams/{-$pgNum}': typeof TeamsChar123PgNumChar125Route
   '/account': typeof AccountIndexRoute
   '/district/$districtAbbreviation/insights': typeof DistrictDistrictAbbreviationInsightsRoute
-  '/district/$districtAbbreviation/stats': typeof DistrictDistrictAbbreviationStatsRoute
   '/district/$districtAbbreviation/{-$year}': typeof DistrictDistrictAbbreviationChar123YearChar125Route
   '/events/$districtAbbreviation/$year': typeof EventsDistrictAbbreviationYearRoute
   '/suggest/review/$suggestionType': typeof SuggestReviewSuggestionTypeRoute
@@ -304,7 +287,6 @@ export interface FileRoutesByTo {
   '/team/$teamNumber/stats': typeof TeamTeamNumberStatsRoute
   '/team/$teamNumber/{-$year}': typeof TeamTeamNumberChar123YearChar125Route
   '/suggest/review': typeof SuggestReviewIndexRoute
-  '/district/$districtAbbreviation/champs/$year': typeof DistrictDistrictAbbreviationChampsYearRoute
 }
 export interface FileRoutesById {
   __root__: typeof rootRouteImport
@@ -333,7 +315,6 @@ export interface FileRoutesById {
   '/teams/{-$pgNum}': typeof TeamsChar123PgNumChar125Route
   '/account/': typeof AccountIndexRoute
   '/district/$districtAbbreviation/insights': typeof DistrictDistrictAbbreviationInsightsRoute
-  '/district/$districtAbbreviation/stats': typeof DistrictDistrictAbbreviationStatsRoute
   '/district/$districtAbbreviation/{-$year}': typeof DistrictDistrictAbbreviationChar123YearChar125Route
   '/events/$districtAbbreviation/$year': typeof EventsDistrictAbbreviationYearRoute
   '/suggest/review/$suggestionType': typeof SuggestReviewSuggestionTypeRoute
@@ -342,7 +323,6 @@ export interface FileRoutesById {
   '/team/$teamNumber/stats': typeof TeamTeamNumberStatsRoute
   '/team/$teamNumber/{-$year}': typeof TeamTeamNumberChar123YearChar125Route
   '/suggest/review/': typeof SuggestReviewIndexRoute
-  '/district/$districtAbbreviation/champs/$year': typeof DistrictDistrictAbbreviationChampsYearRoute
 }
 export interface FileRouteTypes {
   fileRoutesByFullPath: FileRoutesByFullPath
@@ -372,7 +352,6 @@ export interface FileRouteTypes {
     | '/teams/{-$pgNum}'
     | '/account/'
     | '/district/$districtAbbreviation/insights'
-    | '/district/$districtAbbreviation/stats'
     | '/district/$districtAbbreviation/{-$year}'
     | '/events/$districtAbbreviation/$year'
     | '/suggest/review/$suggestionType'
@@ -381,7 +360,6 @@ export interface FileRouteTypes {
     | '/team/$teamNumber/stats'
     | '/team/$teamNumber/{-$year}'
     | '/suggest/review/'
-    | '/district/$districtAbbreviation/champs/$year'
   fileRoutesByTo: FileRoutesByTo
   to:
     | '/'
@@ -409,7 +387,6 @@ export interface FileRouteTypes {
     | '/teams/{-$pgNum}'
     | '/account'
     | '/district/$districtAbbreviation/insights'
-    | '/district/$districtAbbreviation/stats'
     | '/district/$districtAbbreviation/{-$year}'
     | '/events/$districtAbbreviation/$year'
     | '/suggest/review/$suggestionType'
@@ -418,7 +395,6 @@ export interface FileRouteTypes {
     | '/team/$teamNumber/stats'
     | '/team/$teamNumber/{-$year}'
     | '/suggest/review'
-    | '/district/$districtAbbreviation/champs/$year'
   id:
     | '__root__'
     | '/'
@@ -446,7 +422,6 @@ export interface FileRouteTypes {
     | '/teams/{-$pgNum}'
     | '/account/'
     | '/district/$districtAbbreviation/insights'
-    | '/district/$districtAbbreviation/stats'
     | '/district/$districtAbbreviation/{-$year}'
     | '/events/$districtAbbreviation/$year'
     | '/suggest/review/$suggestionType'
@@ -455,7 +430,6 @@ export interface FileRouteTypes {
     | '/team/$teamNumber/stats'
     | '/team/$teamNumber/{-$year}'
     | '/suggest/review/'
-    | '/district/$districtAbbreviation/champs/$year'
   fileRoutesById: FileRoutesById
 }
 export interface RootRouteChildren {
@@ -483,7 +457,6 @@ export interface RootRouteChildren {
   TeamsChar123PgNumChar125Route: typeof TeamsChar123PgNumChar125Route
   AccountIndexRoute: typeof AccountIndexRoute
   DistrictDistrictAbbreviationInsightsRoute: typeof DistrictDistrictAbbreviationInsightsRoute
-  DistrictDistrictAbbreviationStatsRoute: typeof DistrictDistrictAbbreviationStatsRoute
   DistrictDistrictAbbreviationChar123YearChar125Route: typeof DistrictDistrictAbbreviationChar123YearChar125Route
   EventsDistrictAbbreviationYearRoute: typeof EventsDistrictAbbreviationYearRoute
   SuggestReviewSuggestionTypeRoute: typeof SuggestReviewSuggestionTypeRoute
@@ -492,7 +465,6 @@ export interface RootRouteChildren {
   TeamTeamNumberStatsRoute: typeof TeamTeamNumberStatsRoute
   TeamTeamNumberChar123YearChar125Route: typeof TeamTeamNumberChar123YearChar125Route
   SuggestReviewIndexRoute: typeof SuggestReviewIndexRoute
-  DistrictDistrictAbbreviationChampsYearRoute: typeof DistrictDistrictAbbreviationChampsYearRoute
 }
 
 declare module '@tanstack/react-router' {
@@ -672,13 +644,6 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof DistrictDistrictAbbreviationInsightsRouteImport
       parentRoute: typeof rootRouteImport
     }
-    '/district/$districtAbbreviation/stats': {
-      id: '/district/$districtAbbreviation/stats'
-      path: '/district/$districtAbbreviation/stats'
-      fullPath: '/district/$districtAbbreviation/stats'
-      preLoaderRoute: typeof DistrictDistrictAbbreviationStatsRouteImport
-      parentRoute: typeof rootRouteImport
-    }
     '/district/$districtAbbreviation/{-$year}': {
       id: '/district/$districtAbbreviation/{-$year}'
       path: '/district/$districtAbbreviation/{-$year}'
@@ -735,13 +700,6 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof TeamTeamNumberChar123YearChar125RouteImport
       parentRoute: typeof rootRouteImport
     }
-    '/district/$districtAbbreviation/champs/$year': {
-      id: '/district/$districtAbbreviation/champs/$year'
-      path: '/district/$districtAbbreviation/champs/$year'
-      fullPath: '/district/$districtAbbreviation/champs/$year'
-      preLoaderRoute: typeof DistrictDistrictAbbreviationChampsYearRouteImport
-      parentRoute: typeof rootRouteImport
-    }
   }
 }
 
@@ -782,8 +740,6 @@ const rootRouteChildren: RootRouteChildren = {
   AccountIndexRoute: AccountIndexRoute,
   DistrictDistrictAbbreviationInsightsRoute:
     DistrictDistrictAbbreviationInsightsRoute,
-  DistrictDistrictAbbreviationStatsRoute:
-    DistrictDistrictAbbreviationStatsRoute,
   DistrictDistrictAbbreviationChar123YearChar125Route:
     DistrictDistrictAbbreviationChar123YearChar125Route,
   EventsDistrictAbbreviationYearRoute: EventsDistrictAbbreviationYearRoute,
@@ -793,8 +749,6 @@ const rootRouteChildren: RootRouteChildren = {
   TeamTeamNumberStatsRoute: TeamTeamNumberStatsRoute,
   TeamTeamNumberChar123YearChar125Route: TeamTeamNumberChar123YearChar125Route,
   SuggestReviewIndexRoute: SuggestReviewIndexRoute,
-  DistrictDistrictAbbreviationChampsYearRoute:
-    DistrictDistrictAbbreviationChampsYearRoute,
 }
 export const routeTree = rootRouteImport
   ._addFileChildren(rootRouteChildren)
