+++ pwa/tests/district-insights.spec.ts
@@ -1,28 +1,31 @@
 import { expect, test } from '@playwright/test';
 
-test('district insights team table sorts on header click', async ({ page }) => {
-  await page.goto('/district/fim/insights');
+test.describe('/district/fim/insights', () => {
+  test.beforeEach(async ({ page }) => {
+    await page.goto('/district/fim/insights');
+    await page.locator('body[data-hydrated]').waitFor();
+  });
+
+  test('shows the all-time district insights heading', async ({ page }) => {
+    await expect(
+      page.getByRole('heading', { name: 'FIRST in Michigan Insights' }),
+    ).toBeVisible();
+  });
+
+  test('shows event leaderboards from the former stats page', async ({
+    page,
+  }) => {
+    await page.getByRole('tab', { name: 'Events' }).click();
+
+    await expect(page.getByText('Most District Seasons')).toBeVisible();
+  });
+});
+
+test('retired district stats URL is not found', async ({ page }) => {
+  await page.goto('/district/fim/stats');
   await page.locator('body[data-hydrated]').waitFor();
 
-  await page.getByRole('tab', { name: 'Team Data' }).click();
-
-  const teamHeader = page.getByRole('columnheader').first().getByRole('button');
-  await expect(teamHeader).toBeVisible();
-  await expect(teamHeader).toHaveAccessibleName(/Team/);
-  await expect(teamHeader).toHaveAttribute('title', 'Sort ascending');
-
-  const teamCells = page.getByRole('row').locator('td:first-child');
-  await expect(teamCells.first()).toHaveText(/^\d+$/);
-  const lowestTeam = Number(await teamCells.first().textContent());
-
-  await teamHeader.click();
-  await expect(teamHeader).toHaveAttribute('title', 'Sort descending');
-  await expect(teamCells.first()).toHaveText(String(lowestTeam));
-
-  await teamHeader.click();
-  await expect(teamHeader).toHaveAttribute('title', 'Clear sort');
-  await expect(teamCells.first()).toHaveText(/^\d+$/);
-  expect(Number(await teamCells.first().textContent())).toBeGreaterThan(
-    lowestTeam,
-  );
+  await expect(
+    page.getByRole('heading', { name: 'Error 404 - Page Not Found' }),
+  ).toBeVisible();
 });
