+++ pwa/tests/district.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-test('past-year district championship page stops polling', async ({ page }) => {
+test('past-year district championship tab stops polling', async ({ page }) => {
   await page.clock.install();
 
   const requestUrls: string[] = [];
@@ -9,8 +9,9 @@ test('past-year district championship page stops polling', async ({ page }) => {
     request.url().includes('/api/v3/events/2019'),
   );
 
-  await page.goto('/district/fim/champs/2019');
+  await page.goto('/district/fim/2019');
   await page.locator('body[data-hydrated]').waitFor();
+  await page.getByRole('tab', { name: 'Champs' }).click();
   await firstEventsRequest;
 
   const eventsRequestCount = requestUrls.filter((url) =>
@@ -24,11 +25,53 @@ test('past-year district championship page stops polling', async ({ page }) => {
   ).toHaveLength(eventsRequestCount);
 });
 
+test.describe('/district/fim/2024 header', () => {
+  test.beforeEach(async ({ page }) => {
+    await page.goto('/district/fim/2024');
+    await page.locator('body[data-hydrated]').waitFor();
+  });
+
+  [['Rankings'], ['Events'], ['Teams'], ['Champs']].forEach(([tabName]) => {
+    test(`shows the ${tabName} tab`, async ({ page }) => {
+      await expect(page.getByRole('tab', { name: tabName })).toBeVisible();
+    });
+  });
+
+  [['511 teams'], ['27 events']].forEach(([summary]) => {
+    test(`shows ${summary} below the title`, async ({ page }) => {
+      await expect(page.getByText(summary, { exact: true })).toBeVisible();
+    });
+  });
+
+  test('links to the District Championship below the title', async ({
+    page,
+  }) => {
+    await expect(
+      page.getByRole('link', { name: 'District Championship' }),
+    ).toHaveAttribute('href', '/event/2024micmp');
+  });
+
+  test('links to all-time insights from the year selector', async ({
+    page,
+  }) => {
+    await page.getByRole('button', { name: '2024' }).click();
+
+    await expect(
+      page.getByRole('menuitem', { name: 'Insights' }),
+    ).toHaveAttribute('href', '/district/fim/insights');
+  });
+
+  test('shows the championship tracker in the Champs tab', async ({ page }) => {
+    await page.getByRole('tab', { name: 'Champs' }).click();
+
+    await expect(page.getByText('All Rankings')).toBeVisible();
+  });
+});
+
 test.describe('/district/fim/2024 rankings', () => {
   test.beforeEach(async ({ page }) => {
     await page.goto('/district/fim/2024');
     await page.locator('body[data-hydrated]').waitFor();
-    await page.getByRole('tab', { name: 'Rankings' }).click();
   });
 
   test('shows the Advancement column', async ({ page }) => {
